bunny_lords/requirements.txt
pygame-ce>=2.4.0
pygame-gui>=0.6.9
